+++ src/ui/control/terrain_control.ts
@@ -66,7 +66,7 @@ export default class TerrainControl implements IControl {
     _updateTerrainIcon() {
         this._terrainButton.classList.remove('maplibregl-ctrl-terrain');
         this._terrainButton.classList.remove('maplibregl-ctrl-terrain-enabled');
-        if (this._map.style.terrain) {
+        if (this._map.terrain) {
             this._terrainButton.classList.add('maplibregl-ctrl-terrain-enabled');
             this._terrainButton.title = this._map._getUIString('TerrainControl.disableTerrain');
         } else {